Tài 18/03/30
Một số hiệu chỉnh tổng quát bài toán

+++ AutoRebaring/Database/LevelTitle.cs
@@ -11,7 +11,7 @@ namespace AutoRebaring.Database
     {
         public long ID { get; set; }
 
-        public long ProjectID { get; set; }
+        public int ProjectID { get; set; }
 
         [Required]
         [StringLength(50)]
